create base module
