+++ pyriemann/__init__.py
@@ -1,4 +1,5 @@
 from ._version import __version__
+from . import base
 from . import classification
 from . import tangentspace
 from . import channelselection
@@ -13,6 +14,7 @@ from . import datasets
 
 __all__ = [
     '__version__',
+    'base',
     'classification',
     'tangentspace',
     'channelselection',
